+++ sketch/pulse.js
@@ -1,8 +1,8 @@
 class Pulse extends Program {
-  constructor(lightArray, interval) {
-    super(lightArray, interval);
+  constructor(lightMatrix, interval) {
+    super(lightMatrix, interval);
 
-    this.currentLightIndex = 0;
+    this.currentRowIndex = 0;
 
     this.fadeTime = 370;
   }
@@ -15,10 +15,32 @@ class Pulse extends Program {
 
       if (this._isTimeForUpdate()) {
         this._fadeCurrent();
-        this._next();
+        this._nextRow();
         this._lightUpCurrent();
         this.lastChange = millis();
       }
+    }
+  }
+
+  _fadeCurrent() {
+    this._fadeRow(this.currentRowIndex, this.fadeTime);
+  }
+
+  _lightUpCurrent() {
+    this._lightUpRow(this.currentRowIndex, this.defaultLightColor);
+  }
+
+  _nextRow() {
+    this.currentRowIndex++;
+    if (this.currentRowIndex >= this.lights.rows) {
+      this.currentRowIndex = 0;
+    }
+  }
+
+  _previousRow() {
+    this.currentRowIndex--;
+    if (this.currentRowIndex < 0) {
+      this.currentRowIndex = this.lights.rows -1;
     }
   }