+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 - Lights handle their own fading.
 - Light coordination separated into programs, e.g. Pulse, Reverse Pulse.
+- Lights organised as matrix instead of array.
 
 ### Changed
 - Project name: Light Array -> Lights. To accommodate matrix configuration.